Render image on click.
