+++ web/hep_evd.js
@@ -3,13 +3,12 @@
 //
 
 import * as THREE from "three";
-import { Lut } from "three/addons/math/Lut.js";
 import Stats from "three/addons/libs/stats.module.js";
 
 import { THEME, applyConfig } from "./constants.js";
 import { getData } from "./data_loader.js";
 import { RenderState } from "./render_state.js";
-import { animate, onWindowResize } from "./rendering.js";
+import { animate, onWindowResize, renderImage } from "./rendering.js";
 import { nextState, previousState, updateStateUI } from "./states.js";
 import {
   dragElement,
@@ -24,7 +23,6 @@ import {
   toggleTheme,
 } from "./ui.js";
 import { highlightParticleOnMouseMove } from "./interactions.js";
-import { addColourMap, getContinuousLutConf } from "./colourmaps.js";
 
 // Set off the data loading straight away.
 // For big events, this can take a while, so we want to do it in parallel with
@@ -36,7 +34,7 @@ const threeDCamera = new THREE.PerspectiveCamera(
   50,
   window.innerWidth / window.innerHeight,
   0.1,
-  1e6
+  1e6,
 );
 const twoDCamera = new THREE.OrthographicCamera(
   window.innerWidth / -2,
@@ -44,7 +42,7 @@ const twoDCamera = new THREE.OrthographicCamera(
   window.innerHeight / 2,
   window.innerHeight / -2,
   -1,
-  1e6
+  1e6,
 );
 const renderer = new THREE.WebGLRenderer({
   alpha: true,
@@ -85,7 +83,7 @@ const threeDRenderer = new RenderState(
   mcHits.filter((hit) => hit.position.dim === "3D"),
   markers.filter((marker) => marker.position.dim === "3D"),
   detectorGeometry,
-  stateInfo
+  stateInfo,
 );
 const twoDRenderer = new RenderState(
   "2D",
@@ -96,7 +94,7 @@ const twoDRenderer = new RenderState(
   mcHits.filter((hit) => hit.position.dim === "2D"),
   markers.filter((marker) => marker.position.dim === "2D"),
   detectorGeometry,
-  stateInfo
+  stateInfo,
 );
 threeDRenderer.otherRenderer = twoDRenderer;
 twoDRenderer.otherRenderer = threeDRenderer;
@@ -152,7 +150,7 @@ window.addEventListener(
     onWindowResize(threeDRenderer, renderer);
     onWindowResize(twoDRenderer, renderer);
   },
-  false
+  false,
 );
 document.resetView = () => {
   threeDRenderer.resetView();
@@ -168,75 +166,6 @@ canvas.addEventListener("mousemove", (event) => {
   currentlyHighlighting = highlightParticleOnMouseMove(
     renderStates,
     currentlyHighlighting,
-    event
+    event,
   );
 });
-
-fetch("/images")
-  .then((response) => response.json())
-  .then((images) => {
-    images.forEach((image) => {
-      const dataValues = [...new Set(image.data.flat())];
-      const maxValue = Math.max(...dataValues);
-      const minValue = Math.min(...dataValues);
-
-      const canvas = document.createElement("canvas");
-      const ctx = canvas.getContext("2d");
-      canvas.width = image.width;
-      canvas.height = image.height;
-      const imageData = ctx.createImageData(image.width, image.height);
-
-      const hitColours = image.data;
-
-      if (image.imageType === "Monochrome") {
-        const colourLut = new Lut("cooltowarm", 10);
-        const lutConfig = getContinuousLutConf();
-        addColourMap(colourLut, lutConfig.name, lutConfig.size);
-
-        colourLut.setMax(maxValue);
-        colourLut.setMin(minValue);
-
-        for (let row = 0; row < image.height; row++) {
-          for (let col = 0; col < image.width; col++) {
-            if (hitColours[row][col] !== 0)
-              hitColours[row][col] = colourLut.getColor(hitColours[row][col]);
-          }
-        }
-      }
-
-      // Check if its a THREE.Color and scale between 0 - 255.
-      // Finally if its just a raw RGB value, just use it.
-      const getColour = (colour) => {
-        if (colour instanceof THREE.Color) {
-            return [255 * colour.r, 255 * colour.g, 255 * colour.b, 255];
-        } else if (Array.isArray(colour) && colour.length === 3) {
-            return [...colour, 255];
-        } else {
-            return [0.0, 0.0, 0.0, 0.0];
-        }
-      };
-
-      for (let i = 0; i < imageData.data.length; i += 4) {
-        const row = Math.floor(i / (image.height * 4));
-        const col = Math.floor(i / 4) % image.width;
-        const colour = getColour(hitColours[row][col]);
-
-        // If the value isn't 0, scale every channel between 0 - 255, inverted.
-        imageData.data[i + 0] = colour[0];
-        imageData.data[i + 1] = colour[1];
-        imageData.data[i + 2] = colour[2];
-        imageData.data[i + 3] = colour[3];
-      }
-
-      ctx.putImageData(imageData, 0, 0);
-
-      const im = new Image();
-      im.src = canvas.toDataURL("image/png");
-
-      const imDiv = document.createElement("div");
-      imDiv.classList.add("evd_image");
-      imDiv.appendChild(im);
-      dragElement(imDiv);
-      document.body.appendChild(imDiv);
-    });
-  });
