+++ web/ui.js
@@ -7,6 +7,7 @@ import * as THREE from "three";
 import { COLOUR_MAPS, DEFAULT_MAPS } from "./colourmaps.js";
 import { BUTTON_ID, GITHUB_URL, TO_THEME } from "./constants.js";
 import { createParticleMenu } from "./particle_menu.js";
+import { renderImage } from "./rendering.js";
 
 /**
  * Populates a dropdown menu with buttons based on the given hit property map.
@@ -91,7 +92,7 @@ export function populateTypeToggle(hitDim, hitTypesMap, onClick = (_) => {}) {
       "btn-outline",
       "btn-accent",
       "m-1",
-      "nohover",
+      "nohover"
     );
     newButton.style.textTransform = "capitalize";
     newButton.innerText = entry;
@@ -114,7 +115,7 @@ export function populateMarkerToggle(
   hitDim,
   markers,
   particles,
-  onClick = (_) => {},
+  onClick = (_) => {}
 ) {
   // Get the div to populate, and clear it to start.
   const classDiv = document.getElementById(`markers_${hitDim}`);
@@ -143,7 +144,7 @@ export function populateMarkerToggle(
       "btn-outline",
       "btn-accent",
       "m-1",
-      "nohover",
+      "nohover"
     );
     newButton.style.textTransform = "capitalize";
     newButton.innerText = entry;
@@ -217,7 +218,7 @@ export function enableInteractionTypeToggle(hitType, particles, onClick) {
       "btn-outline",
       "btn-accent",
       "m-1",
-      "nohover",
+      "nohover"
     );
     newButton.innerText = interactionType;
     newButton.id = `particles_${hitType}_${interactionType}`;
@@ -290,7 +291,7 @@ export function setupParticleMenu(renderState) {
   const onClickAction = (particle) => {
     const particleID = particle.id;
     const particleMenuEntry = document.getElementById(
-      `particle_${particleID}_${renderState.hitDim}`,
+      `particle_${particleID}_${renderState.hitDim}`
     );
     const label = particleMenuEntry.querySelector("span");
 
@@ -308,7 +309,7 @@ export function setupParticleMenu(renderState) {
   createParticleMenu(
     renderState.hitDim,
     renderState.particleData.particleMap,
-    onClickAction,
+    onClickAction
   );
 }
 
@@ -381,7 +382,7 @@ export function screenshotEvd(renderer) {
   const contentType = "image/jpeg";
 
   const byteCharacters = atob(
-    imageData.substr(`data:${contentType};base64,`.length),
+    imageData.substr(`data:${contentType};base64,`.length)
   );
   const bytes = [];
 
@@ -509,7 +510,7 @@ export function saveState(states) {
 
     if (name === undefined || name === "") return;
     const visibleState = Array.from(states.values()).find(
-      (state) => state.visible,
+      (state) => state.visible
     );
     const store = window.localStorage;
 
@@ -550,8 +551,8 @@ export function saveState(states) {
         closed = true;
         cleanUp();
       },
-      { once: true },
-    ),
+      { once: true }
+    )
   );
 
   inputSave.addEventListener(
@@ -562,7 +563,7 @@ export function saveState(states) {
       doSave();
       cleanUp();
     },
-    { once: true },
+    { once: true }
   );
 }
 
@@ -573,7 +574,7 @@ export function saveState(states) {
  */
 export function loadState(renderStates) {
   const visibleState = Array.from(renderStates.values()).find(
-    (state) => state.visible,
+    (state) => state.visible
   );
   const store = window.localStorage;
 
@@ -592,7 +593,7 @@ export function loadState(renderStates) {
   }
 
   const validSaveStates = saveStates.filter(
-    (state) => state.hitDim === visibleState.hitDim,
+    (state) => state.hitDim === visibleState.hitDim
   );
 
   if (validSaveStates === null) return;
@@ -633,8 +634,8 @@ export function loadState(renderStates) {
         closed = true;
         cleanUp();
       },
-      { once: true },
-    ),
+      { once: true }
+    )
   );
 
   selectButton.addEventListener(
@@ -645,7 +646,7 @@ export function loadState(renderStates) {
       cleanUp();
       visibleState.triggerEvent("change");
     },
-    { once: true },
+    { once: true }
   );
 }
 
@@ -654,7 +655,7 @@ export function loadState(renderStates) {
  */
 export function pickColourscheme(states) {
   const visibleState = Array.from(states.values()).find(
-    (state) => state.visible,
+    (state) => state.visible
   );
   const store = window.localStorage;
 
@@ -688,7 +689,7 @@ export function pickColourscheme(states) {
       option.text = csName;
       categoricalSelect.add(option);
       continuousSelect.add(option.cloneNode(true));
-    },
+    }
   );
 
   // Finally show the modal.
@@ -737,8 +738,8 @@ export function pickColourscheme(states) {
         closed = true;
         cleanUp();
       },
-      { once: true },
-    ),
+      { once: true }
+    )
   );
 
   selectButton.addEventListener(
@@ -750,7 +751,7 @@ export function pickColourscheme(states) {
 
       visibleState.triggerEvent("fullUpdate");
     },
-    { once: true },
+    { once: true }
   );
 }
 
@@ -912,27 +913,37 @@ export function setupMobileUI(renderer) {
  * @param {Array} images - The images to populate the dropdown with.
  */
 export function populateImages(images) {
-
+  const dropDownButton = document.getElementById(`image_dropdown_button`);
   const dropDown = document.getElementById(`image_dropdown`);
   dropDown.innerHTML = "";
 
-  if (images.length == 0) {
-    dropDown.style.display = "none";
+  if (images.length === 0) {
+    dropDownButton.style.display = "none";
     return;
   } else {
-    dropDown.style.display = "block";
+    dropDownButton.style.display = "grid";
   }
 
   const showImage = (image) => {
-    console.log(`Showing image: ${image.name}`);
-  }
+    console.log(`Showing image: ${image.label}`);
+
+    const im = renderImage(image);
+
+    const imDiv = document.createElement("div");
+    imDiv.classList.add("evd_image");
+    imDiv.appendChild(im);
+    dragElement(imDiv);
+    document.body.appendChild(imDiv);
+
+    console.log(imDiv);
+  };
 
   images.forEach((image) => {
     const listElement = document.createElement("li");
     const newButton = document.createElement("li");
     newButton.style.textTransform = "capitalize";
-    newButton.innerText = image.name;
-    newButton.id = `${image.name}`;
+    newButton.innerText = image.label;
+    newButton.id = `${image.label}`;
     newButton.addEventListener("click", () => showImage(image));
     listElement.appendChild(newButton);
     dropDown.appendChild(listElement);
